Move physics enter/exit back to reactors
