+++ packages/spatial/src/transform/components/EntityTree.tsx
@@ -426,7 +426,7 @@ export function useAncestorWithComponent(entity: Entity, component: ComponentTyp
       unmounted = true
       root.stop()
     }
-  }, [])
+  }, [entity, component])
 
   return result.value
 }
@@ -468,7 +468,7 @@ export function useChildWithComponent(entity: Entity, component: ComponentType<a
       unmounted = true
       root.stop()
     }
-  }, [])
+  }, [entity, component])
 
   return result.value
 }
